pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Your name (or your organization/company/team)" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "beartype>=0.22.9",
    "einops>=0.8.2",
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "jaxtyping>=0.3.9",
    "lmdb>=1.7.5",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "mkdocs",
    "pandas>=3.0.2",
    "pip",
    "python-dotenv",
    "ruff",
    "scipy>=1.17.1",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "wandb>=0.25.1",
]
requires-python = "~=3.11.0"


[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting
ignore = ["F722"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
